Moved storage manager to a separate file
